Fix imports from refactor
